feat: pokemon, refactor, improved logging
