+++ package.json
@@ -43,8 +43,8 @@
     "example:basic": "npx tsx examples/basic/main.ts",
     "example:weather": "npx tsx examples/weather/main.ts",
     "example:custom": "npx tsx examples/custom/main.ts",
-    "ai": "npx tsx src/ai/agent.ts",
-    "ai:test": "npx tsx src/ai/test.ts"
+    "ai:triage-weather-email": "npx tsx src/ai/test.ts",
+    "ai:pokemon": "npx tsx src/ai/example/pokemon.ts"
   },
   "dependencies": {
     "@ai-sdk/openai": "^0.0.70",
